Split test cases to different files
