+++ tests/data/query_router/schema/query_ds.json
